# src/main/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSubscriberRequest": {
        "description": "Webhook subscriber registration request",
        "properties": {
          "events": {
            "description": "Event filter; the single element * subscribes to every event. Omitted or empty means every event.",
            "example": [
              "marketplace.snapshot.approved",
              "marketplace.snapshot.rejected"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "description": "Gateway-local subscriber name",
            "example": "ci-bot",
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "type": "string"
          },
          "url": {
            "description": "Target URL; scheme must be on the configured allowlist (default http/https)",
            "example": "https://ci.example.com/hooks/skills-gateway",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreatedSubscriber": {
        "description": "A freshly created subscriber; the only time the signing secret is ever returned",
        "properties": {
          "createdAt": {
            "description": "Creation time",
            "format": "date-time",
            "type": "string"
          },
          "events": {
            "description": "Event filter; the single element * means every event",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "description": "Subscriber id",
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "description": "Subscriber name",
            "type": "string"
          },
          "secret": {
            "description": "Signing secret - shown exactly once",
            "type": "string"
          },
          "url": {
            "description": "Target URL",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Finding": {
        "description": "One thing a vetter found in a snapshot",
        "properties": {
          "id": {
            "description": "Stable rule identifier, e.g. aws-access-key-id",
            "example": "aws-access-key-id",
            "type": "string"
          },
          "location": {
            "description": "Where in the snapshot it was found, normally path:line",
            "type": "string"
          },
          "message": {
            "description": "Reviewer-facing explanation; never echoes the matched secret",
            "type": "string"
          },
          "severity": {
            "description": "How much the finding matters",
            "enum": [
              "info",
              "low",
              "medium",
              "high",
              "critical"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Run": {
        "description": "One execution of the vetting chain against a snapshot",
        "properties": {
          "chain": {
            "description": "Identity of the chain that produced the run: vetter@version in chain order",
            "example": "prompt-injection@1,secret-scan@1",
            "type": "string"
          },
          "finishedAt": {
            "description": "When the run finished, or null if it never did",
            "format": "date-time",
            "type": "string"
          },
          "outcome": {
            "description": "Fail-closed aggregate of the run's verdicts",
            "enum": [
              "clear",
              "clear_with_waivers",
              "blocked"
            ],
            "type": "string"
          },
          "runId": {
            "description": "Chain run id",
            "format": "int64",
            "type": "integer"
          },
          "snapshotId": {
            "description": "Snapshot the run vetted",
            "format": "int64",
            "type": "integer"
          },
          "startedAt": {
            "description": "When the run started",
            "format": "date-time",
            "type": "string"
          },
          "trigger": {
            "description": "What caused the run",
            "enum": [
              "ingestion",
              "revet-scheduled",
              "revet-manual"
            ],
            "type": "string"
          },
          "verdicts": {
            "description": "The run's verdicts, in chain order",
            "items": {
              "$ref": "#/components/schemas/VerdictView"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "SubscriberView": {
        "description": "A webhook subscriber without its signing secret",
        "properties": {
          "createdAt": {
            "description": "Creation time",
            "format": "date-time",
            "type": "string"
          },
          "enabled": {
            "description": "Whether deliveries are queued for this subscriber",
            "type": "boolean"
          },
          "events": {
            "description": "Event filter; the single element * means every event",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "description": "Subscriber id",
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "description": "Subscriber name",
            "type": "string"
          },
          "url": {
            "description": "Target URL",
            "type": "string"
          }
        },
        "type": "object"
      },
      "Suppression": {
        "description": "A finding an active waiver is currently suppressing",
        "properties": {
          "approvedBy": {
            "description": "Identity that accepted the risk",
            "type": "string"
          },
          "expiresAt": {
            "description": "When the acceptance lapses",
            "format": "date-time",
            "type": "string"
          },
          "location": {
            "description": "Where the finding was located",
            "type": "string"
          },
          "ruleId": {
            "description": "Finding rule identifier",
            "type": "string"
          },
          "vetter": {
            "description": "Vetter whose verdict carried the finding",
            "type": "string"
          },
          "waiverId": {
            "description": "Waiver suppressing it",
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UncoveredFinding": {
        "description": "A blocking finding that no active waiver covers",
        "properties": {
          "location": {
            "description": "Where the finding was located",
            "type": "string"
          },
          "message": {
            "description": "Reviewer-facing explanation",
            "type": "string"
          },
          "ruleId": {
            "description": "Finding rule identifier",
            "type": "string"
          },
          "severity": {
            "description": "How much it matters",
            "enum": [
              "info",
              "low",
              "medium",
              "high",
              "critical"
            ],
            "type": "string"
          },
          "vetter": {
            "description": "Vetter whose verdict carried the finding",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "VerdictView": {
        "description": "One vetter's recorded verdict within a chain run",
        "properties": {
          "detail": {
            "description": "One-line summary: how many findings and the worst severity, or — for a clean pass with no findings — what the vetter examined",
            "type": "string"
          },
          "findings": {
            "description": "What the vetter found",
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array"
          },
          "position": {
            "description": "Position of the vetter in the chain",
            "format": "int32",
            "type": "integer"
          },
          "reportUrl": {
            "description": "External report URL, when the vetter produced one",
            "type": "string"
          },
          "state": {
            "description": "The vetter's conclusion",
            "enum": [
              "pass",
              "warn",
              "fail",
              "error",
              "pending",
              "disabled",
              "not_reached"
            ],
            "type": "string"
          },
          "verdictId": {
            "description": "Verdict id",
            "format": "int64",
            "type": "integer"
          },
          "vetter": {
            "description": "Vetter name",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "VetterView": {
        "description": "A vetter configured in the vetting chain",
        "properties": {
          "description": {
            "description": "What the vetter looks for, and what it cannot see",
            "type": "string"
          },
          "external": {
            "description": "Whether the verdict is delegated to an operator-configured external service, so that a verdict the gateway did not reach itself never reads as one that it did",
            "type": "boolean"
          },
          "name": {
            "description": "Stable vetter name",
            "type": "string"
          },
          "order": {
            "description": "Position in the chain",
            "format": "int32",
            "type": "integer"
          },
          "version": {
            "description": "Identity of the rule set the vetter currently carries (GW_VETTING_0012)",
            "type": "string"
          }
        },
        "type": "object"
      },
      "VettingOverrideRecord": {
        "description": "An administrator's override of a blocked vetting outcome on a snapshot",
        "properties": {
          "blockingVetters": {
            "description": "The vetters that were blocking",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "description": "Override id",
            "format": "int64",
            "type": "integer"
          },
          "overriddenAt": {
            "description": "When the override was recorded",
            "format": "date-time",
            "type": "string"
          },
          "overriddenBy": {
            "description": "Identity of the administrator who overrode the block",
            "type": "string"
          },
          "reason": {
            "description": "The administrator's stated reason",
            "type": "string"
          },
          "snapshotId": {
            "description": "The snapshot approved over the block",
            "format": "int64",
            "type": "integer"
          },
          "uncoveredFindings": {
            "description": "Summary of the blocking findings no waiver covered",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "VettingView": {
        "description": "A snapshot's latest vetting chain run, the waivers over it, and the chain that produced it",
        "properties": {
          "outcome": {
            "description": "The effective outcome, which is what gates approval: the run's verdicts with every waived finding removed. `clear_with_waivers` means nothing objects any more only because an active waiver is suppressing a finding. A snapshot with no run is blocked.",
            "enum": [
              "clear",
              "clear_with_waivers",
              "blocked"
            ],
            "type": "string"
          },
          "override": {
            "$ref": "#/components/schemas/VettingOverrideRecord",
            "description": "Present when an administrator approved this snapshot over a blocked vetting outcome (GW_VETTING_0028); its presence is what surfaces the override so it is never indistinguishable from a clean approval. Null otherwise."
          },
          "recordedOutcome": {
            "description": "What the vetters themselves concluded, before any waiver was applied",
            "enum": [
              "clear",
              "blocked"
            ],
            "type": "string"
          },
          "run": {
            "$ref": "#/components/schemas/Run",
            "description": "The latest run with its verdicts and findings, or null if the chain never ran"
          },
          "snapshotId": {
            "description": "Snapshot id",
            "format": "int64",
            "type": "integer"
          },
          "suppressed": {
            "description": "Findings an active waiver is currently suppressing, keyed by vetter, rule and location",
            "items": {
              "$ref": "#/components/schemas/Suppression"
            },
            "type": "array"
          },
          "uncovered": {
            "description": "Blocking findings that no active waiver covers; the waivers approval still needs",
            "items": {
              "$ref": "#/components/schemas/UncoveredFinding"
            },
            "type": "array"
          },
          "vetters": {
            "description": "The vetters configured in the chain, in the order they run",
            "items": {
              "$ref": "#/components/schemas/VetterView"
            },
            "type": "array"
          },
          "waivers": {
            "description": "Waivers of this snapshot's marketplace whose rule appears in this run, active and lapsed alike, so an expired acceptance stays visible",
            "items": {
              "$ref": "#/components/schemas/WaiverView"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "WaiverView": {
        "description": "A waiver as returned by the API, with its activity resolved as of now",
        "properties": {
          "active": {
            "description": "Whether the waiver suppresses anything right now",
            "type": "boolean"
          },
          "approvedBy": {
            "description": "Identity that accepted the risk",
            "type": "string"
          },
          "createdAt": {
            "description": "When the waiver was created",
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "description": "When the acceptance lapses; never null",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Waiver id",
            "format": "int64",
            "type": "integer"
          },
          "justification": {
            "description": "Why the risk is accepted",
            "type": "string"
          },
          "marketplace": {
            "description": "Marketplace the waiver belongs to",
            "type": "string"
          },
          "revokedAt": {
            "description": "When the waiver was revoked, or null",
            "format": "date-time",
            "type": "string"
          },
          "revokedBy": {
            "description": "Identity that revoked it, or null",
            "type": "string"
          },
          "ruleId": {
            "description": "Finding rule identifier being accepted",
            "example": "aws-access-key-id",
            "type": "string"
          },
          "scope": {
            "description": "How the scope value is matched",
            "enum": [
              "snapshot",
              "path"
            ],
            "type": "string"
          },
          "scopeValue": {
            "description": "A commit SHA for snapshot scope, a repository-relative path for path scope",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
...
  "info": {
    "contact": {
      "name": "jimisola",
      "url": "https://github.com/skillsgateway/skillsgateway"
    },
...
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Skills Gateway",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/snapshots/{id}/vetting": {
      "get": {
        "description": "The snapshot's latest vetting chain run: each vetter's verdict in chain order, the findings behind it, the waivers currently suppressing any of them, and the fail-closed effective aggregate that gates approval. A snapshot the chain has never run against reports a blocked outcome and no run.",
        "operationId": "vetting",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VettingView"
                }
              }
            },
            "description": "The latest chain run and the configured chain"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VettingView"
                }
              }
            },
            "description": "Snapshot not found"
          }
        },
        "summary": "Snapshot vetting verdicts",
        "tags": [
          "Vetting"
        ]
      }
    },
...
    "/api/v1/webhooks": {
      "get": {
        "description": "Every registered receiver with its event filter. Signing secrets are never returned.",
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubscriberView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List webhook subscribers",
        "tags": [
          "Webhooks"
        ]
      },
      "post": {
        "description": "Registers a receiver for marketplace lifecycle events. The signing secret is returned exactly once, in this response, and is never readable afterwards.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSubscriber"
                }
              }
            },
            "description": "Subscriber registered; response carries the show-once secret"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSubscriber"
                }
              }
            },
            "description": "Disallowed URL scheme, or an unknown event name"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSubscriber"
                }
              }
            },
            "description": "A subscriber with that name already exists"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSubscriber"
                }
              }
            },
            "description": "Invalid subscriber name"
          }
        },
        "summary": "Register a webhook subscriber",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
    "/api/v1/webhooks/{id}": {
      "delete": {
        "description": "Removes the subscriber and its delivery history; no further events are queued for it.",
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Subscriber deleted"
          },
          "404": {
            "description": "Subscriber not found"
          }
        },
        "summary": "Delete a webhook subscriber",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
    "/hooks/{marketplace}": {
      "post": {
        "description": "Trigger endpoint for a forge push webhook. Authenticated solely by an HMAC-SHA256 signature of the raw request body (GitHub-compatible X-Hub-Signature-256 header) against the secret generated when the marketplace's sync mode was set to webhook. The payload is ignored: a valid signature only causes the registered upstream URL's default branch to be ingested into a held quarantine snapshot, asynchronously.",
        "operationId": "trigger",
        "parameters": [
          {
            "in": "path",
            "name": "marketplace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Hub-Signature-256",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Trigger accepted; ingestion queued"
          },
          "403": {
            "description": "Missing or invalid signature"
          },
          "404": {
            "description": "Unknown marketplace, or its sync mode is not webhook"
          },
          "413": {
            "description": "Request body exceeds the configured bound"
          }
        },
        "summary": "Forge push webhook",
        "tags": [
          "Sync"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}